requirements.txt
ruff
mypy
pytest
fastapi
uvicorn[standard]
httpx
python-dotenv
claude-agent-sdk